Drop programs reducer
