--- website/frontend/src/store/actions/actions.ts
@@ -1,169 +1,0 @@
-import { CreateType, GearKeyring } from '@gear-js/api';
-import { ProgramActionTypes, ProgramModel, ProgramPaginationModel } from 'types/program';
-import { programService } from 'services/ProgramsRequestService';
-import ServerRPCRequestService from 'services/ServerRPCRequestService';
-import { RPC_METHODS, LOCAL_STORAGE } from 'consts';
-import { PaginationModel, UserPrograms } from 'types/common';
-import { getLocalPrograms, getLocalProgram, getLocalProgramMeta, isDevChain } from 'helpers';
-import { nodeApi } from '../../api/initApi';
-import { AlertTypes } from 'types/alerts';
-import { ReactNode } from 'react';
-import { AlertCustomOptionsWithType } from 'react-alert';
-
-const fetchUserProgramsAction = () => ({ type: ProgramActionTypes.FETCH_USER_PROGRAMS });
-const fetchUserProgramsSuccessAction = (payload: ProgramPaginationModel) => ({
-  type: ProgramActionTypes.FETCH_USER_PROGRAMS_SUCCESS,
-  payload,
-});
-const fetchAllProgramsSuccessAction = (payload: ProgramPaginationModel) => ({
-  type: ProgramActionTypes.FETCH_ALL_PROGRAMS_SUCCESS,
-  payload,
-});
-const fetchUserProgramsErrorAction = () => ({ type: ProgramActionTypes.FETCH_USER_PROGRAMS_ERROR });
-
-const fetchProgramAction = () => ({ type: ProgramActionTypes.FETCH_PROGRAM });
-const fetchProgramSuccessAction = (payload: ProgramModel) => ({
-  type: ProgramActionTypes.FETCH_PROGRAM_SUCCESS,
-  payload,
-});
-export const resetProgramAction = () => ({ type: ProgramActionTypes.RESET_PROGRAM });
-
-const fetchProgramErrorAction = () => ({ type: ProgramActionTypes.FETCH_PROGRAM_ERROR });
-
-export const programUploadStartAction = () => ({ type: ProgramActionTypes.PROGRAM_UPLOAD_START });
-export const programUploadResetAction = () => ({ type: ProgramActionTypes.PROGRAM_UPLOAD_RESET });
-export const programUploadSuccessAction = () => ({ type: ProgramActionTypes.PROGRAM_UPLOAD_SUCCESS });
-export const programUploadFailedAction = (payload: string) => ({
-  type: ProgramActionTypes.PROGRAM_UPLOAD_FAILED,
-  payload,
-});
-
-export const fetchProgramPayloadTypeAction = (payload: string) => ({
-  type: ProgramActionTypes.FETCH_PROGRAM_PAYLOAD_TYPE,
-  payload,
-});
-export const resetProgramPayloadTypeAction = () => ({ type: ProgramActionTypes.RESET_PROGRAM_PAYLOAD_TYPE });
-
-export const sendMessageStartAction = () => ({ type: ProgramActionTypes.SEND_MESSAGE_START });
-export const sendMessageResetAction = () => ({ type: ProgramActionTypes.SEND_MESSAGE_RESET });
-export const sendMessageSuccessAction = () => ({ type: ProgramActionTypes.SEND_MESSAGE_SUCCESS });
-export const sendMessageFailedAction = (payload: string) => ({ type: ProgramActionTypes.SEND_MESSAGE_FAILED, payload });
-
-export const uploadMetaStartAction = () => ({ type: ProgramActionTypes.META_UPLOAD_START });
-export const uploadMetaResetAction = () => ({ type: ProgramActionTypes.META_UPLOAD_RESET });
-const uploadMetaSuccessAction = () => ({ type: ProgramActionTypes.META_UPLOAD_SUCCESS });
-const uploadMetaFailedAction = (payload: string) => ({ type: ProgramActionTypes.META_UPLOAD_FAILED, payload });
-
-export const programStatusAction = (payload: string) => ({ type: ProgramActionTypes.PROGRAM_STATUS, payload });
-
-export const fetchGasAction = (payload: number) => ({ type: ProgramActionTypes.FETCH_GAS, payload });
-export const resetGasAction = () => ({ type: ProgramActionTypes.RESET_GAS });
-
-export const getUserProgramsAction = (params: UserPrograms) => (dispatch: any) => {
-  const getPrograms = isDevChain() ? getLocalPrograms : programService.fetchUserPrograms;
-
-  dispatch(fetchUserProgramsAction());
-  getPrograms(params)
-    .then((data) => {
-      dispatch(fetchUserProgramsSuccessAction(data.result));
-    })
-    .catch(() => dispatch(fetchUserProgramsErrorAction()));
-};
-
-export const getAllProgramsAction = (params: PaginationModel) => (dispatch: any) => {
-  const getPrograms = isDevChain() ? getLocalPrograms : programService.fetchAllPrograms;
-
-  dispatch(fetchUserProgramsAction());
-  getPrograms(params)
-    .then((data) => {
-      dispatch(fetchAllProgramsSuccessAction(data.result));
-    })
-    .catch(() => dispatch(fetchUserProgramsErrorAction()));
-};
-
-export const getProgramAction = (id: string) => (dispatch: any) => {
-  const getProgram = isDevChain() ? getLocalProgram : programService.fetchProgram;
-
-  dispatch(fetchProgramAction());
-  getProgram(id)
-    .then((data) => {
-      dispatch(fetchProgramSuccessAction(data.result));
-    })
-    .catch(() => dispatch(fetchProgramErrorAction()));
-};
-
-export const handleProgramError = (error: string) => (dispatch: any, getState: any) => {
-  const { programs } = getState();
-  if (programs.isProgramUploading) {
-    dispatch(programUploadFailedAction(error));
-  } else if (programs.isMessageSending) {
-    dispatch(sendMessageFailedAction(error));
-  } else if (programs.isMetaUploading) {
-    dispatch(uploadMetaFailedAction(error));
-  }
-};
-
-export const handleProgramSuccess = () => (dispatch: any, getState: any) => {
-  const { programs } = getState();
-  if (programs.isProgramUploading) {
-    dispatch(programUploadSuccessAction());
-  } else if (programs.isMessageSending) {
-    dispatch(sendMessageSuccessAction());
-  } else if (programs.isMetaUploading) {
-    dispatch(uploadMetaSuccessAction());
-  }
-};
-
-export const subscribeToEvents = (showAlert: (message?: ReactNode, options?: AlertCustomOptionsWithType) => void) => {
-  const filterKey = localStorage.getItem(LOCAL_STORAGE.PUBLIC_KEY_RAW);
-  nodeApi.subscribeToProgramEvents(({ method, data: { info, reason } }) => {
-    // @ts-ignore
-    if (info.origin.toHex() === filterKey) {
-      showAlert(`${method}\n ${info.programId.toHex()}`, { type: reason ? AlertTypes.ERROR : AlertTypes.SUCCESS });
-    }
-  });
-
-  nodeApi.subscribeToLogEvents(async ({ data: { source, dest, reply, payload } }) => {
-    let meta = null;
-    let decodedPayload: any;
-    const programId = source.toHex();
-    const apiRequest = new ServerRPCRequestService();
-
-    const { result } = isDevChain()
-      ? await getLocalProgramMeta(programId)
-      : await apiRequest.getResource(RPC_METHODS.GET_METADATA, { programId });
-
-    if (result && result.meta) {
-      meta = JSON.parse(result.meta);
-    } else {
-      showAlert('Metadata is not added', { type: AlertTypes.ERROR });
-    }
-
-    try {
-      decodedPayload =
-        meta.output && !(reply.isSome && reply.unwrap()[1].toNumber() !== 0)
-          ? CreateType.decode(meta.output, payload, meta).toHuman()
-          : payload.toHuman();
-    } catch (error) {
-      console.error('Decode payload failed');
-    }
-    // @ts-ignore
-    if (dest.toHex() === filterKey) {
-      // TODO: add payload parsing
-      const msg = `LOG from program\n ${source.toHex()}\n ${decodedPayload ? `Response: ${decodedPayload}` : ''}`;
-      const options = {
-        type:
-          (reply.isSome && reply.unwrap()[1].toNumber() === 0) || reply.isNone ? AlertTypes.SUCCESS : AlertTypes.ERROR,
-      };
-
-      showAlert(msg, options);
-    }
-  });
-
-  nodeApi.subscribeToTransferEvents(({ data: { from, to, value } }) => {
-    if (to.toHex() === filterKey) {
-      const msg = `TRANSFER BALANCE\n FROM:${GearKeyring.encodeAddress(from.toHex())}\n VALUE:${value.toString()}`;
-      showAlert(msg, { type: AlertTypes.INFO });
-    }
-  });
-};
