--- website/frontend/src/store/index.ts
@@ -1,10 +1,0 @@
-import { applyMiddleware, createStore } from 'redux';
-import thunk from 'redux-thunk';
-// eslint-disable-next-line import/no-extraneous-dependencies
-import { composeWithDevTools } from 'redux-devtools-extension';
-
-import { rootReducer } from './reducers';
-
-const store = createStore(rootReducer, composeWithDevTools(applyMiddleware(thunk)));
-
-export default store;
